Implementing basic version ECDF

+++ StatsNET/Statistic.Basic.UnitTests/StatisticHelperTests.cs
@@ -17,7 +17,7 @@ namespace Statistic.Basic.UnitTests
 
         //28, 35, 42, 90, 70, 56, 75, 66, 30, 89, 75, 64, 81, 69, 55, 83, 72, 68, 73, 16
         [TestCase("1,3,5",0,3,2)]
-        public void GetAbsoluteFrequencyShouldReturnTheCorrectFrequencyForASpecificInterval(string data, double startInterval, double endInterval, double expected)
+        public void GetAbsoluteFrequencyShouldReturnTheCorrectFrequencyForASpecificInterval(string data, double startInterval, double endInterval, int expected)
         {
             GivenASetOfData(data);
             GivenASpecificRange(startInterval, endInterval);
